pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pixellab-tool"
version = "0.1.0"
description = "CLI tool for PixelLab pixel art generation API"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "HandyCommunication"}
]
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "python-dotenv>=1.0",
    "Pillow>=9.0",
    "rich>=13.0",
]

[project.scripts]
pixellab = "pixellab_tool.cli:cli"

[project.urls]
Homepage = "https://github.com/HandyCommunication/PixelLab-Tool"

[tool.setuptools.packages.find]
include = ["pixellab_tool*"]
